requirements.txt
numpy>=1.24
scipy>=1.10
networkx>=3.1
matplotlib>=3.7
PyYAML>=6.0
